Respell cursor_bnode as cursor_node; finished restoring offset return from tux_create_entry; add ERR_CAST to userspace ERR_PTR header

+++ user/kernel/tux3.h
@@ -646,7 +646,7 @@ struct btree new_btree(struct sb *sb, struct btree_ops *ops);
 
 /* dir.c */
 tux_dirent *tux_find_entry(struct inode *dir, const char *name, int len, struct buffer_head **result);
-int tux_create_entry(struct inode *dir, const char *name, int len, inum_t inum, unsigned mode);
+loff_t tux_create_entry(struct inode *dir, const char *name, int len, inum_t inum, unsigned mode);
 int tux_delete_entry(struct buffer_head *buffer, tux_dirent *entry);
 int tux_readdir(struct file *file, void *state, filldir_t filldir);
 extern const struct file_operations tux_dir_fops;
